Create a thread-safe, multi-threaded queue-processing service. The service should read from an input Queue of Strings and carry out an operation on them.
Assume that the operation is so processor-intensive that you’ll need to use parallel threads to keep up with the rate at which items are added to the queue. (Assume the service will eventually be deployed on a machine with many cores.)

Implement two operations, both of which assume that the strings are gigantic integers (e.g. 100+ digits):
Maintain a count of all those that are larger than some cut-off value.
Similarly, find the maximum and minimum values.

The service should provide a thread-safe way for other threads to query the current best answer (either the count so far, or the maximum/minimum values encountered so far).

Write a utility that will pump numbers into your queue from a (possibly very large) text file.